Refactor gyration methods into Gyration class
